requirements.txt
absl-py==0.7.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
bokeh==1.3.4
boto==2.49.0
boto3==1.9.184
botocore==1.12.146
bz2file==0.98
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
colorama==0.3.9
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
docutils==0.14
entrypoints==0.3
gast==0.2.2
grpcio==1.16.1
h5py==2.9.0
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
jedi==0.14.0
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.3.0
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.2
olefile==0.46
packaging==19.1
pandas==0.24.2
pandas-profiling==1.4.2
pandocfilters==1.4.2
parso==0.5.0
PDPbox==0.2.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.15.3
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
pyzmq==18.0.2
qtconsole==4.5.1
requests==2.22.0
rope==0.14.0
rsa==3.4.2
s3transfer==0.2.1
scikit-learn==0.21.3
scikit-plot==0.3.7
scipy==1.3.0
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
smart-open==1.8.4
SQLAlchemy==1.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.31.1
traitlets==4.3.2
urllib3==1.24.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wrapt==1.11.2
xlrd==1.2.0
py-xgboost==0.90
